Tighten up types and initialization for private properties

+++ src/VersionUpdater/Downloader/CurlDownloader.php
@@ -12,13 +12,14 @@
 namespace Psy\VersionUpdater\Downloader;
 
 use Psy\Exception\ErrorException;
+use Psy\Exception\RuntimeException;
 use Psy\Shell;
 use Psy\VersionUpdater\Downloader;
 
 class CurlDownloader implements Downloader
 {
-    private $tempDir = null;
-    private $outputFile = null;
+    private ?string $tempDir = null;
+    private ?string $outputFile = null;
 
     /** {@inheritDoc} */
     public function setTempDir(string $tempDir)
@@ -71,13 +72,17 @@ class CurlDownloader implements Downloader
     /** {@inheritDoc} */
     public function getFilename(): string
     {
+        if ($this->outputFile === null) {
+            throw new RuntimeException('Call download() first');
+        }
+
         return $this->outputFile;
     }
 
     /** {@inheritDoc} */
     public function cleanup()
     {
-        if (\file_exists($this->outputFile)) {
+        if ($this->outputFile !== null && \file_exists($this->outputFile)) {
             \unlink($this->outputFile);
         }
     }
